// src/api/people/people.interface.ts
import { BackSetFields, ID } from 'src/common/common.interface';
import { IFile } from 'src/file/file.interface';
import { IFilm } from '../film/film.interface';
import { IPlanet } from '../planet/planet.interface';
import { ISpecie } from '../specie/specie.interface';
import { IStarship } from '../starship/starship.interface';
import { IVehicle } from '../vehicle/vehicle.interface';

export interface IPeople extends BackSetFields, DPeople, RPeople, ID {}
export interface DPeople {
  name: string;
  mass: string;
  height: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
}
export interface RPeople {
  films: IFilm[];
  species: ISpecie[];
  vehicles: IVehicle[];
  starships: IStarship[];
  homeworld: IPlanet;
  files: IFile[];
}
export interface SPeople {
  films: string[];
  species: string[];
  vehicles: string[];
  starships: string[];
  homeworld: string;
}
export interface RIDPeople {
  films: ID[];
  species: ID[];
  vehicles: ID[];
  starships: ID[];
  homeworld: ID;
  files: ID[];
}